+++ src/pictures/filter.js
@@ -4,6 +4,8 @@ module.exports = {
   getFilteredPictures: function(picturesArray, filter) {
     var picturesToFilter = picturesArray.slice(0);
 
+
+
     switch (filter) {
       case 'filter-popular':
         break;
